requirements.txt
flask
requests
aiohttp
protobuf
pycryptodome
PyJWT
PyYAML
python-dotenv